Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
-    float x,y,z;
+    int x,y,z;
     scanf("%f %f %f",&x,%y,%z);
-    printf("Average: %.2f",(x+y+z)/3.00);
+    printf("Average: %.2f",(x+y+z)/3);
     return 0;
 }
